+++ selfdrive/car/hyundai/radar_interface.py
@@ -11,7 +11,7 @@ RADAR_START_ADDR = 0x500
 RADAR_MSG_COUNT = 32
 def get_radar_can_parser(CP):
 
-  if False: #Params().get_bool("NewRadarInterface"):
+  if Params().get_bool("NewRadarInterface"):
 
     signals = []
     checks = []
@@ -46,7 +46,7 @@ def get_radar_can_parser(CP):
 class RadarInterface(RadarInterfaceBase):
   def __init__(self, CP):
     super().__init__(CP)
-    self.new_radar = False #Params().get_bool("NewRadarInterface")
+    self.new_radar = Params().get_bool("NewRadarInterface")
     self.updated_messages = set()
     self.trigger_msg = 0x420 if not self.new_radar else RADAR_START_ADDR + RADAR_MSG_COUNT - 1
     self.track_id = 0